Change adapters module structure
Turn adapters.py file into separate module with submodules.
